pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ssl-ctc-finetuning"
version = "0.1.0"
description = "Config-driven SSL-CTC fine-tuning framework (XLSR-first defaults)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "datasets==2.21.0",
  "librosa==0.11.0",
  "numpy==2.2.6",
  "soundfile==0.13.1",
  "tokenizers==0.22.2",
  "torch>=2.8.0",
  "torchaudio>=2.8.0",
  "transformers==4.57.6",
  "jiwer==4.0.0",
  "evaluate==0.4.6",
  "accelerate==1.11.0",
  "matplotlib==3.10.1",
  "PyYAML==6.0.2",
  "cn2an==0.5.23",
  "num2words==0.5.14",
]

[project.scripts]
ctc-train = "ctc_framework.cli.train:main"
ctc-build-vocab = "ctc_framework.cli.build_vocab:main"
ctc-eval = "ctc_framework.cli.eval:main"
ctc-plot-compare = "ctc_framework.cli.plot_compare:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
